modules/location_service/location_service_add/requirements.txt
kafka-python 
psycopg2-binary
SQLAlchemy
